// components/student/studentMyProject/studentMyProject/studentMyProjectListStyle.scss.d.ts
export type Styles = {
  'buttonAction': string;
  'buttonActionDelete': string;
  'buttonActionEdit': string;
  'buttonsActionsContainer': string;
  'imgContainer': string;
  'itemContainer': string;
  'itemContentContainer': string;
  'itemGrid': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;